refactor(estimation): centralize the Cholesky-Ω packing convention in parameterization
The column-major lower-triangle packing that `pack_params` is the authority for was
re-derived independently in four places. Moved into `estimation/parameterization.rs`
(next to `pack_params`) as the single source of truth — pure code motion + integer/f64
algebra, no result reordered:

- `lower_tri_entries`, `chol_pack`, `block_chol_full` moved verbatim out of
  sens_outer_gradient.rs (its bare-name call sites resolve via the extended import).
- `chol_lt_idx` moved verbatim out of api.rs.
- agq.rs now reaches `chol_pack` at `parameterization::` instead of the cross-module
  `sens_outer_gradient::` path.
- api.rs's `se_kappa` block offset and types.rs's `omega_se_at` block offset now call
  the shared `chol_lt_idx` instead of re-inlining `i*n - i*(i-1)/2` (the `omega_se_at`
  call is guarded `r < n_eta` to preserve None-on-out-of-range and keep the
  `debug_assert` from firing). Bit-identical index for every valid input.

The larger `PackedLayout`/`omega_packed_len` abstraction from the design is intentionally
deferred: its only consumers are the optional extract_standard_errors/packed_param_label
rewrites, so adding it now would be dead code.

New Tier-1 parity tests pin `lower_tri_entries` frozen order (n=1..3) and that
`chol_lt_idx` round-trips its position for n=1..4. Verified: cargo test --lib =>
2512 passed, 0 failed; the SE index-map anchors (test_se_kappa_block_n3_column_major_indexing,
test_se_omega_block_*) stay green.
